services/delivery-service/requirements.txt
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
aio-pika==9.4.1
